requirements.txt
altgraph==0.17.4
blinker==1.9.0
buildozer==1.5.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distlib==0.3.9
docutils==0.21.2
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
ntplib==0.4.0
packaging==25.0
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pefile==2023.2.7
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
ptyprocess==0.7.0
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyperclip==1.9.0
pypiwin32==223
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
requests==2.32.3
screeninfo==0.8.1
setuptools==80.7.1
sh==2.2.2
typing_extensions==4.13.2
urllib3==2.4.0
ursina==7.0.0
virtualenv==20.31.2
Werkzeug==3.1.3
